Feat(Bonus): 보너스 숫자를 보관할 클래스 생성
